Renaming and updating documentation
